requirements.txt
django==3.2.9
psycopg2-binary==2.9.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0